moved type around for offchainvote
